Fix EdStem #1718 and related
The new password hashing part of the project introduced a logical
race condition between the client and the server: the client might
send requests while the server hasn't finished setting up the account
for a client, and the requests (which are handled in a different
thread) would fail, causing spurious test failures under some thread
schedules.

This change makes the tests wait for 100ms before issuing requests,
giving the server enough time to set up the account.